added respective changes for a new method getItem()

+++ src/main/java/com/cash/dtos/CatalogueItemRequestDto.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class CatalogueItemRequestDto {
     private String title;
     private String description;
-    private double startingPrice;
+    private int startingPrice;
     private int durationHours;
     private int sellerId;
 }
